fixed flashing rendering

+++ src/src/main/java/BeatKeeper.java
@@ -17,7 +17,7 @@ public class BeatKeeper {
     }
 
     // Various getters and setters
-    public int getTime() { return time; }
+    public float getTime() { return time; }
 
     public void setTime(int time) { this.time = time; }
